pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = "The backend server for the ineview services"
license = "MIT"
authors = []

[tool.poetry.dependencies]
Django = "^4.0.2"
django-cors-headers = "^3.11.0"
djangorestframework = "^3.13.1"
python = ">=3.10,<3.11"
tensorflow = "^2.8.0"
opencv-contrib-python = "^4.5.5"
tensorflow-gpu = "^2.8.0"
fer = "^22.0.1"
ffmpeg-python = "^0.2.0"
schedule = "^1.1.0"
django-extensions = "^3.1.5"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
